fix(police): wire the foreign-resize detector, and four rule fixes
The one that matters: `snapshotGridMoved` shipped fully plumbed — kaval through
padi-client, unit-tested — and NOTHING called it. `attach.ts`'s header claimed
the multi-client gap was closed and even named the intended call site, while the
app still painted the corruption. That is the same overclaim the debate caught
in `snapshotAnswersGrid`, repeated one layer up: capability shipped, fix not.

`Terminal.tsx` now refuses a frame whose SERVED grid moved, in the same
recoverable channel as a locally stale one. Two viewers then ping-pong the
width, which is the multi-client contract's own stated outcome — "contention
costs a repaint, never the attach loop" — bounded by the reopen backoff. The
header names its real consumer rather than an intended one.

Rule fixes:
- `padi`'s `grid?: { cols; rows }` was an anonymous twin of the `EndpointGrid`
  the same files already use for `resizeTo`. Reused.
- kaval labelled the new `grid` "contract 5.3", already taken by the
  meaningful-output edges — ambiguous by construction in the exact place this
  code is otherwise careful. Now 5.4.
- The three git marks moved out of `PrPip.tsx` into `./forgeIcons`, matching the
  convention this PR itself enforced when it emptied them from
  `client/src/ui/Icons.tsx`. One consumer is a fact about today, not a reason to
  inline.
- The hydrate script's `chmod … || true` was the one swallowed error with no
  justification beside `exit 1` everywhere else. It is best-effort because the
  `rm -rf` after it is the step that must succeed and reports its own failure —
  now said, rather than assumed.
